update experiment code
