+++ catkin_ws/src/doorgym/src/inference_vx300s.py
@@ -10,7 +10,7 @@ import rospy
 import tf
 from tf.transformations import euler_from_quaternion, quaternion_from_euler
 
-sys.path.append(os.getcwd())
+sys.path.append("../DoorGym")
 import a2c_ppo_acktr
 
 from std_srvs.srv import Trigger, TriggerRequest
@@ -27,12 +27,14 @@ class Inference:
         self.get_knob_srv = rospy.ServiceProxy("/gazebo/get_link_state", GetLinkState)
         self.go_joint_srv = rospy.ServiceProxy("/robot/go_joint_pose", joint_pose)
         self.get_door_angle_srv = rospy.ServiceProxy("/gazebo/get_joint_properties", GetJointProperties)
+        self.ran = rospy.ServiceProxy("husky_ur5/random", Trigger)
+        self.arm_go_home = rospy.ServiceProxy("/robot/ur5/go_home", Trigger)
         self.joint = np.zeros(23)
         self.dis = 0
         self.listener = tf.TransformListener()
         self.joint_value = [0.0, 0.0, 0.0, 0.0, 0.0, 0.0] 
 
-        load_name = "trained_models/ppo/doorenv-v0_push-load.125.pt"
+        load_name = "../DoorGym/model/husky_ur5_push.pt"
         self.actor_critic, ob_rms = torch.load(load_name)
         self.actor_critic = self.actor_critic.eval()
         self.actor_critic.to("cuda:0")
@@ -40,6 +42,8 @@ class Inference:
         self.recurrent_hidden_states = torch.zeros(1, self.actor_critic.recurrent_hidden_state_size)
         self.masks = torch.zeros(1, 1)
         
+        self.arm_go_home()
+        self.ran()
         self.inference()
 
     def joint_state_cb(self, msg):
@@ -89,7 +93,7 @@ class Inference:
         pos = self.get_knob_srv(req)
 
         try:
-            trans, _ = self.listener.lookupTransform("/map", "vx300s/ee_gripper_link", rospy.Time(0))
+            trans, _ = self.listener.lookupTransform("/world", "vx300s/ee_gripper_link", rospy.Time(0))
         except (tf.LookupException, tf.ConnectivityException, tf.ExtrapolationException) as e:
             print("Service call failed: %s"%e)
 
@@ -117,12 +121,6 @@ class Inference:
             req.joint_name = "hinge_door_0::hinge"
 
             res = self.get_door_angle_srv(req)
-
-            # push parameter
-            if res.position[0] <= -0.45:
-                joint_action[2] *= 8
-                joint_action[6] *= -5
-                joint_action[4] *= -0.2
 
             # vx300s push paramter
             self.joint_value[0] += joint_action[2] * -0.009
@@ -153,6 +151,6 @@ class Inference:
         print("time", end - begin)
 
 if __name__ == '__main__':
-    rospy.init_node("doorgym_node", anonymous=False)
+    rospy.init_node("husky_vx300s_node", anonymous=False)
     inference = Inference()
     rospy.spin()
